Added levls and new font

+++ CircleClicker.py
@@ -15,10 +15,19 @@ def run_game_events(mx, my):
         if event.type == QUIT:
             pygame.quit()
 
-        if event.type == MOUSEBUTTONDOWN and circle.mouse_collide(mx, my) and not timer.is_finished():
+        if event.type == MOUSEBUTTONDOWN and circle.mouse_collide(mx, my) and timer.is_started() and not timer.is_finished():
             score += 1
             circle.delete()
-            circle = Circle(screen, (255, 0, 0), 50)
+            circle = Circle(screen, (255, 0, 0), None)
+            if level == "EASY":
+                circle.set_radius(50)
+
+            elif level == "MEDIUM":
+                circle.set_radius(30)
+
+            elif level == "HARD":
+                circle.set_radius(10)
+
             circle.draw_circle()
 
 
@@ -26,18 +35,44 @@ start_game = False
 
 
 def main():
-    global start_game
+    global start_game, level
     run = True
     while run:
-
         mx, my = pygame.mouse.get_pos()
-        screen.blit(level_surf, level_surf_rect)
-        medium_btn.draw(screen)
-        easy_btn.draw(screen)
-        hard_btn.draw(screen)
 
         if not start_game:
             screen.blit(title_text.get_text(), title_text.get_text_rect())
+
+            screen.blit(level_surf, level_surf_rect)
+
+            if medium_btn.draw(screen):
+                level = "MEDIUM"
+                circle.set_radius(20)
+                medium_btn.update(GREEN, BLUE)
+                if easy_btn.is_selected():
+                    easy_btn.update(GREEN, BLUE)
+                if hard_btn.is_selected():
+                    hard_btn.update(GREEN, BLUE)
+
+            if easy_btn.draw(screen):
+                level = "EASY"
+                circle.set_radius(20)
+                easy_btn.update(GREEN, BLUE)
+                if medium_btn.is_selected():
+                    medium_btn.update(GREEN, BLUE)
+                if hard_btn.is_selected():
+                    hard_btn.update(GREEN, BLUE)
+
+            if hard_btn.draw(screen):
+                level = "HARD"
+                hard_btn.update(GREEN, BLUE)
+                if easy_btn.is_selected():
+                    easy_btn.update(GREEN, BLUE)
+                if medium_btn.is_selected():
+                    medium_btn.update(GREEN, BLUE)
+
+            hard_btn.draw(screen)
+
             if start_btn.draw(screen):
                 start_game = True
 
@@ -49,13 +84,22 @@ def main():
 
         else:
             timer.start()
-            timer_text = Text(30, 20, timer.get_time(), "assets/arcade_font.TTF", 32, RED, None)
+            timer_text = Text(30, 20, timer.get_time(), FONT, 32, RED, None)
 
-            score_text = Text(550, 20, f"Score: {score}", "assets/arcade_font.TTF", 32, (0, 0, 255))
+            score_text = Text(550, 20, f"Score: {score}", FONT, 32, GREEN)
 
             screen.fill((0, 0, 0))
             screen.blit(score_text.get_text(), score_text.get_text_rect())
             screen.blit(timer_text.get_text(), timer_text.get_text_rect())
+
+            if level == "EASY":
+                circle.set_radius(50)
+
+            elif level == "MEDIUM":
+                circle.set_radius(30)
+
+            elif level == "HARD":
+                circle.set_radius(10)
 
             circle.draw_circle()
